Add proper check for dotty presentation compiler, refactor
